train logreg with toggle button
