+++ tools/sense_studio/utils.py
@@ -1,6 +1,10 @@
 import json
+import numpy as np
 import os
 import torch
+
+from joblib import dump
+from sklearn.linear_model import LogisticRegression
 
 from sense import backbone_networks
 from sense import engine
@@ -96,3 +100,61 @@ def toggle_project_setting(path, setting):
     write_project_config(path, config)
 
     return new_status
+
+
+def train_logreg(path, split, label):
+    """
+    (Re-)Train a logistic regression model on all annotations that have been submitted so far.
+    """
+    tags_dir = os.path.join(path, f"tags_{split}", label)
+    features_dir = os.path.join(path, f"features_{split}", label)
+    logreg_dir = os.path.join(path, 'logreg', label)
+    logreg_path = os.path.join(logreg_dir, 'logreg.joblib')
+
+    annotations = os.listdir(tags_dir)
+    class_weight = {0: 0.5}
+
+    if annotations:
+        features = [os.path.join(features_dir, x.replace('.json', '.npy')) for x in annotations]
+        annotations = [os.path.join(tags_dir, x) for x in annotations]
+        x = []
+        y = []
+
+        for feature in features:
+            feature = np.load(feature)
+
+            for f in feature:
+                x.append(f.mean(axis=(1, 2)))
+
+        for annotation in annotations:
+            with open(annotation, 'r') as f:
+                annotation = json.load(f)['time_annotation']
+
+            pos1 = np.where(np.array(annotation).astype(int) == 1)[0]
+
+            if len(pos1) > 0:
+                class_weight.update({1: 2})
+
+                for p in pos1:
+                    if p + 1 < len(annotation):
+                        annotation[p + 1] = 1
+
+            pos1 = np.where(np.array(annotation).astype(int) == 2)[0]
+
+            if len(pos1) > 0:
+                class_weight.update({2: 2})
+
+                for p in pos1:
+                    if p + 1 < len(annotation):
+                        annotation[p + 1] = 2
+
+            for a in annotation:
+                y.append(a)
+
+        x = np.array(x)
+        y = np.array(y)
+
+        if len(class_weight) > 1:
+            logreg = LogisticRegression(C=0.1, class_weight=class_weight)
+            logreg.fit(x, y)
+            dump(logreg, logreg_path)
